+++ perf/_runner.py
@@ -15,7 +15,7 @@ from perf._cli import (format_run, format_benchmark, format_checks,
 from perf._bench import _load_suite_from_pipe
 from perf._cpu_utils import (format_cpu_list, parse_cpu_list,
                              get_isolated_cpus, set_cpu_affinity)
-from perf._formatter import format_timedelta, format_number, format_sample
+from perf._formatter import format_timedelta, format_number, format_value
 from perf._utils import (MS_WINDOWS, popen_killer, abs_executable,
                          create_environ, create_pipe, WritePipe,
                          get_python_names)
@@ -386,12 +386,12 @@ class Runner:
                 samples.append(value)
 
             if args.verbose:
-                text = format_sample(unit, sample)
+                text = format_value(unit, sample)
                 if is_warmup or is_calibrate:
                     text = ('%s (%s: %s)'
                             % (text,
                                format_number(loops, 'loop'),
-                               format_sample(unit, raw_sample)))
+                               format_value(unit, raw_sample)))
                 print("%s %s: %s" % (sample_name, index, text))
 
             if calibrate and raw_sample < args.min_time: